+++ src/Network/Network/ChangeLog.md
@@ -51,6 +51,7 @@
     - `New/Get/Remove/Set-AzNetworkManagerScopeConnection` 
     - `New/Get/Remove/Set-AzNetworkManagerSubscriptionConnection` 
     - `New/Get/Remove/Set-AzNetworkManagerManagementGroupConnection` 
+* Onboard AgFoodPlatform to Private Link Common Cmdlets
 
 ## Version 4.19.0
 * Updated cmdlets to add new property of `ExplicitProxy` in Azure Firewall Policy.