+++ src/Websites/Websites/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Fixed `Import-AzWebAppKeyVaultCertificate` to use certificate naming convention same as Az-CLI
 
 ## Version 2.11.3
 * Fixed `Publish-AzWebapp` to handle relative paths properly [#18028]